requirements.txt
sqlalchemy==2.0.46
psycopg[binary]==3.3.2
python-dotenv==1.0.0

# Web Framework & Testing
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0

# Data Validation
pydantic==2.5.0

# Password Hashing
bcrypt==4.1.1

# File Upload & Import
python-multipart==0.0.6
openpyxl==3.1.2
